+++ src/cpp/hydra.cpp
@@ -166,21 +166,22 @@ namespace hydra{
 
     }
 
-    floatVector hydraBase::getSubConfiguration( const unsigned int &lowerIndex, const unsigned int &upperIndex ){
+    floatVector hydraBase::getSubConfiguration( const floatMatrix &configurations, const unsigned int &lowerIndex,
+                                                const unsigned int &upperIndex ){
         /*!
          * Get a sub-configuration \f$\bf{F}^{sc}\f$ defined as
          *
          * \f$ F^{sc}_{iI} = F^{\text{lowerIndex}}_{i\hat{I}} F^{\text{lowerIndex} + 1}_{\hat{I}\breve{I}} \cdots F^{\text{upperIndex-1}}_{\bar{I}I} \f$
+         * \param &configurations: The configurations to operate on
          * \param &lowerIndex: The index of the lower configuration (starts at 0 and goes to numConfigurations - 1)
          * \param &upperIndex: The index of the upper configuration (starts at 0 and goes to numConfigurations)
          *   Note, the configuration indicated by the index is NOT included in the sub-configuration
          */
-
-        if ( upperIndex > *getNumConfigurations( ) ){
+        if ( upperIndex > configurations.size( ) ){
 
             std::string message = "The upper index must be less than or equal to the total number of configurations\n";
             message            += "  upperIndex      : " + std::to_string( upperIndex ) + "\n";
-            message            += "  # configurations: " + std::to_string( *getNumConfigurations( ) );
+            message            += "  # configurations: " + std::to_string( configurations.size( ) );
 
             ERROR_TOOLS_CATCH( throw std::runtime_error( message ) );
 
@@ -203,7 +204,7 @@ namespace hydra{
 
         for ( unsigned int i = lowerIndex; i < upperIndex; i++ ){
 
-            Fsc = vectorTools::matrixMultiply( Fsc, ( *getConfigurations( ) )[ i ], ( *dim ), ( *dim ), ( *dim ), ( *dim ) );
+            Fsc = vectorTools::matrixMultiply( Fsc, configurations[ i ], ( *dim ), ( *dim ), ( *dim ), ( *dim ) );
 
         }
 
@@ -211,11 +212,25 @@ namespace hydra{
 
     }
 
+    floatVector hydraBase::getSubConfiguration( const unsigned int &lowerIndex, const unsigned int &upperIndex ){
+        /*!
+         * Get a sub-configuration \f$\bf{F}^{sc}\f$ defined as
+         *
+         * \f$ F^{sc}_{iI} = F^{\text{lowerIndex}}_{i\hat{I}} F^{\text{lowerIndex} + 1}_{\hat{I}\breve{I}} \cdots F^{\text{upperIndex-1}}_{\bar{I}I} \f$
+         * \param &lowerIndex: The index of the lower configuration (starts at 0 and goes to numConfigurations - 1)
+         * \param &upperIndex: The index of the upper configuration (starts at 0 and goes to numConfigurations)
+         *   Note, the configuration indicated by the index is NOT included in the sub-configuration
+         */
+
+        return getSubConfiguration( *getConfigurations( ), lowerIndex, upperIndex );
+
+    }
+
     floatVector hydraBase::getPrecedingConfiguration( const unsigned int &index ){
         /*!
-         * Get the sub-configuration preceding but not including the current index
+         * Get the sub-configuration preceding but not including the index
          * 
-         * \param &index: The index of the current configuration
+         * \param &index: The index of the configuration immediately following the sub-configuration
          */
 
         return getSubConfiguration( 0, index );
@@ -224,12 +239,48 @@ namespace hydra{
 
     floatVector hydraBase::getFollowingConfiguration( const unsigned int &index ){
         /*!
-         * Get the sub configuration following but not including the current index
+         * Get the sub-configuration following but not including the index
          * 
-         * \param &index: The index of the current configuration
+         * \param &index: The index of the current configuration immediately before the sub-configuration
          */
 
         return getSubConfiguration( index + 1, *getNumConfigurations( ) );
+
+    }
+
+    floatVector hydraBase::getPreviousSubConfiguration( const unsigned int &lowerIndex, const unsigned int &upperIndex ){
+        /*!
+         * Get a previous sub-configuration \f$\bf{F}^{sc}\f$ defined as
+         *
+         * \f$ F^{sc}_{iI} = F^{\text{lowerIndex}}_{i\hat{I}} F^{\text{lowerIndex} + 1}_{\hat{I}\breve{I}} \cdots F^{\text{upperIndex-1}}_{\bar{I}I} \f$
+         * \param &lowerIndex: The index of the lower configuration (starts at 0 and goes to numConfigurations - 1)
+         * \param &upperIndex: The index of the upper configuration (starts at 0 and goes to numConfigurations)
+         *   Note, the configuration indicated by the index is NOT included in the sub-configuration
+         */
+
+        return getSubConfiguration( *getPreviousConfigurations( ), lowerIndex, upperIndex );
+
+    }
+
+    floatVector hydraBase::getPreviousPrecedingConfiguration( const unsigned int &index ){
+        /*!
+         * Get the previous sub-configuration preceding but not including the index
+         * 
+         * \param &index: The index of the configuration immediately following the sub-configuration
+         */
+
+        return getPreviousSubConfiguration( 0, index );
+
+    }
+
+    floatVector hydraBase::getPreviousFollowingConfiguration( const unsigned int &index ){
+        /*!
+         * Get the previous sub-configuration following but not including the index
+         * 
+         * \param &index: The index of the current configuration immediately before the sub-configuration
+         */
+
+        return getPreviousSubConfiguration( index + 1, *getNumConfigurations( ) );
 
     }