+++ docs/sphinx/changelog.rst
@@ -22,4 +22,9 @@ Internal Changes
 - Added the decomposition of the incoming state variable vector into the configurations, state variables
   in the non-linear solve, and additional state variables (:merge:`3`). By `Nathan Miller`_.
 - Added function to get a subset of the full deformation gradient (:merge:`6`). By `Nathan Miller`_.
-- Added functions to get the part of the sub-configuration preceding and following a given configuration (:merge:`7`). By `Nathan Miller`_.
+- Added functions to get the part of the sub-configuration preceding and following a given
+  configuration (:merge:`7`). By `Nathan Miller`_.
+- Generalized the computation of the sub-configuration so that one can use either the current or
+  previous configurations (:merge:`8`). By `Nathan Miller`_.
+- Added the computation of the previous sub-configurations and the previous preceding and following
+  sub-configurations given a configuration (:merge:`8`). By `Nathan Miller`_.